Change build so that object files are collected into libhector.a before compiling the final executable.
This change will make it easier to use hector in another code (such as
GCAM).  One consequence of this change is that any code that should
*not* be included when hector is linked into another code
(e.g. main.cpp) should go in the top-level source directory.
Conversely, code that is always necessary for hector to function
(generally everything that isn't main.cpp) must go in the
subdirectories.