Change post-2010 CO2 (anthro & luc) emissions to be driven from code instead of read from file.
All of the non-co2s are still read from input files, but CO2 is pushed
in (starting from 2011) from the subroutines in main.cpp.  The
emissions themselves are grabbed from the RCP45 input file, and the
results are identical to what we get when we parse the RCP45 emissions
at init-time.

The process is still a little warty, since the core doesn't have any
way to look up a component by its inputs.  Therefore, we have to look
up the simpleNbox component by name and send it messages directly.
That required exposing Core::getComponentByName() to the outside
world, which isn't ideal, but we'll get that fixed soon.