Add an implicit conversion to double for unitval type.
Allow us to assign a unitval to a double, stripping off the units in
the process.  This isn't technically necessary for adding the
additional RF reporting; however, it's a small enough change that I
thought I'd piggyback it here.

Also, remove an unnecessary line in unitval::unitval(double,
unit_types).  Also a small change.