Changes to enable and silence warnings in g++.
Most of these were harmless, but one involved assigning a negative
number to an unsigned, which is legal, but it doesn't do what you
think it does. (I.e., it does not produce a value that is unreachable
by starting at 0 and incrementing.)  Since the loop that we use it in
actually starts at 1 (skipping the 0th) column, I adopted 0 as the
flag for "not found".