Time series containing only a single value now acts like a constant
That is, it returns its lone value irrespective of what time was
passed in.  This behavior may seem surprising at first glance, but it
does the right thing for the cases where the situation arises in
practice (namely, during spin-up), and it is a reasonable
generalization of what we mean by a time series.

Fixes #38