Logger will create log directory if it is absent, or fail if unable
On unix, the logger will check to see if the log directory exists and
is writable.  If it doesn't exist, it will try to create it, or die
trying.  Most of the common failure reasons are diagnosed and
reported.

On windows, the logger will attempt to create the log file and will
fail if it can't.  No attempt is made to create a missing directory,
and the error is always reported as "Unable to open log file"

Fixes #39