Add rpath argument to build cmd for executable
Without this argument, Hector will build but will only run if the GSL
library directory is in ld.so.conf.  It turns out that this is the case
on all systems we tested on except Evergreen, and on Evergreen we only
ever tried to run as part of GCAM, which supplies the rpath argument
itself.