Add GPL license
Per discussions with @bpbond, I've added the GPL license to the project.  I went with GPL v.2 partly because it seems to be the default version of the GPL on GitHub (i.e., appears in bold near the top of the list of suggestions) and partly because my brief research on the subject turned up a lot of complaints about v.3.  Honestly, given the nature of this software, I doubt it will make much difference one way or the other.