Allow for CSV filepaths relative to the INI file
CSV filepaths in the INI file can either be full absolute paths or
are considered to be relative to the INI file. This makes it possible
to have an INI file and its associated emissions a self-contained unit,
independent of the location of Hector.

Paths in INI files need to be updated after this change as they were
relative to the path of the Hector executable.