Add reset capability for SimpleNbox component.
The component operates as it always did, and then at the end of each
time step it copies all of the values it's tracking into a time
indexed array.  This proved to be a lot easier to implement than
changing all of the component's working variables into time indexed
values.

This change incidentally fixes a problem with the handling of the
temperature fertilization effect in soil.  This calculation is subject
to a ratchet effect; the calculated value can never be less than the
previous year's value.  However, that previous value was being stored
in a scalar variable.  This scheme worked ok when there was only one
global biome, but would likely have produced incorrect results when we
tried to run with multiple biomes.  The new scheme should be safe to
use with multiple biomes.