When resetting to t < Core::getStartDate(), also rerun spinup.
If you run a scenario, then reset and rerun the spinup, then rerun the
scenario from the new start state, you will get a discrepancy between
the old and new run.  This happens because the initial state produced
by the spinup is slightly different from what the original spinup
produced.  The fractional discrepancy is approximately
1e-3*eps_spinup, where eps_spinup is the spinup tolerance.

With the default eps_spinup of 0.001, these discrepancies are small
enough to be ignorable for scientific purposes, but for things like
Monte Carlo runs we may want to set eps_spinup to something like 1e-6,
so as to ensure that for any given set of parameters results will
agree to within the limits of single precision arithmetic.