Add input files for CO2 constrained scenarios. Add tests for CO2 constraint run.
It turns out we have constraint files only for two of the four RCP
scenarios, so those are the only ones we can add right now.  Then, on
a whim I decided to run one of the other scenarios with the history
(only) constrained, just to see if it would work.  It did, so that
input file is included too.  The rest of the combinations we can add
when we've tracked down the relevant data.