If building R package, use Rcpp's R funs. Otherwise, use Boost.
This relies on the fact that Rcpp defines the `USE_RCPP`
macro (`-DUSE_RCPP`). This allows me to use `#ifdef USE_RCPP` to check
if I'm building the R package (and therefore have access to R
functions and don't need boost) or standalone (so I wouldn't have R
and therefore would need boost).