Add capabilities for adjusted halocarbon forcings
The halocarbon forcings are supplied by the halocarbon components, not
the forcing component, and the former do not know how to correct their
values to be relative to the base year.  To rectify this, we've added
a new set of capabilities for adjusted halocarbon forcings, which are
reported by the forcing component and given relative to the base
year.

The R interface has been changed to report the adjusted forcings
whenever halocarbon forcings are asked for.  There is no provision for
getting the unadjusted halocarbon forcings in R because Hector does
not provide unadjusted forcings for other GHGs, and so there isn't
really anything you can do with the unadjusted halocarbon values.