Update version of Boost on Travis
As Travis upgraded to Ubuntu 16 (xenial), the version and source of
Boost libraries changed. The new default version is 1.58, and it is
not in the core Ubuntu repositories (i.e. does not require a PPA).

Previously, this was causing build errors like the following:

```
Package libboost-filesystem1.55-dev is not available, but is referred to by another package.
This may mean that the package is missing, has been obsoleted, or
is only available from another source
Package libboost-system1.55-dev is not available, but is referred to by another package.
This may mean that the package is missing, has been obsoleted, or
is only available from another source
E: Package 'libboost-system1.55-dev' has no installation candidate
E: Package 'libboost-filesystem1.55-dev' has no installation candidate
```