Use Ubuntu Xenial in Travis build
This fixes an issue whereby there is no
`<std::vector>.erase(<const_iterator>)` method (but there is an
`iterator` method for the same function). This is a known bug in
GCC < 4.9 that has since been resolved. Xenial uses a newer version of
GCC (5.4.0), in which this problem has been resolved.

See this discussion for details:
https://github.com/JGCRI/hector/pull/317#discussion_r298352188