Replace old ubuntu os requests for RCMD tests (#550)
* Switch the OS to ubuntu-latest which seems like a good idea for at least
one of the linux tests.  However, I'm not sure if R 3.6 will always be
available there.  If not we could try on `ubuntu-18.04` but the danger is
always that eventually it will drop out of support too.

* Oops, need to change the "rspm" URL as well.  And while I'm at it update
the comments.

* Switch to explicitly using 20.04 instead of latest at the very least
because otherwise I am not sure what to do with the:
```
Rscript -e 'writeLines(remotes::system_requirements("ubuntu", "20.04"))'
```