Replace custom code to echo to screen with boost.iostreams (#561)
* Replace the custom std::streambuffer in Logger (which was used to provide
the ability to echo logging output to both screen and file) with
boost::iostreams

* We removed include for fstream from the logger header so need to
explicitly include it in test_logger.cpp

* The logger test relied on the fact that we explicitly copied cout's
buffer to capture messages to do the testing.  Boost.iostreams is
keeping reference to the stream itself.  But that is ok.  We can take
the same general approach, we just have to wait to replace the original
cout buffer until TearDown.  And GoogleTest won't be sending messages on
cout so we should be all good.