Alpha and Volscl impact RF output (#745) (#756)
Here we change where the scalers are applied so that the vol, aerosol, and total forcing outputs are updated appropriately. While this PR does not change hector output or behavior between the parameter name change, there are code-breaking changes that have been made to the ini files. The ini files are not backward compatible. 

---------

Co-authored-by: ptrscll <ptrscll@gmail.com>